Changed to match the changes in the SDK

+++ sdk/src/com/podio/sdk/filter/RatingFilter.java
@@ -29,9 +29,9 @@ public class RatingFilter extends BasicPodioFilter {
         super(PATH);
     }
 
-    public RatingFilter withTypeAndId(String ref_type, int ref_id) {
+    public RatingFilter withTypeAndId(String ref_type, long ref_id) {
         addPathSegment(ref_type);
-        addPathSegment(Integer.toString(ref_id, 10));
+        addPathSegment(Long.toString(ref_id, 10));
         return this;
     }
 }
